+++ cmd/server/app/domain/authapp/authapp.proto
