+++ sdk/client/cmd.go
@@ -9,7 +9,7 @@ import (
 // DefaultURL is a convience function for getting a url using the
 // default local model server host:port or pulling from KRONK_WEB_API_HOST.
 func DefaultURL(path string) (string, error) {
-	host := "http://localhost:3000"
+	host := "http://localhost:8080"
 	if v := os.Getenv("KRONK_WEB_API_HOST"); v != "" {
 		host = v
 	}
@@ -25,5 +25,3 @@ func DefaultURL(path string) (string, error) {
 
 	return path, nil
 }
-
-
