+++ sdk/kronk/kronk.go
@@ -21,7 +21,7 @@ import (
 )
 
 // Version contains the current version of the kronk package.
-const Version = "1.7.2"
+const Version = "1.7.3"
 
 // =============================================================================
